requirements.txt
pandas
requests
googe-cloud-storage
google-cloud-pubsub